Removed ExitMenu as it was redundant and added folder for menu items
