+++ UI/Menu/MenuItem.cs
@@ -1,5 +1,4 @@
 ﻿namespace w3_assignment_ksteph.UI;
-
 public class MenuItem
 {
     public string Name { get; set; }
